[release-1.10] Do not require nodemediatedDeviceTypes to prevent broken upgrades (#2534) (#2537)
With #2402 we deprecated MediatedDevicesTypes and
NodeMediatedDeviceTypes.MediatedDevicesTypes for
MediatedDeviceTypes and NodeMediatedDeviceTypes.MediatedDeviceTypes.
We also introduced a mutating webhook to handle the conversion.
Unfortunately make the two new names mandatory is breaking
the validation of the new version of the CRD on OLM upgrades
since only the previous names were there with the old CRD.

We already have func test covering this area,
but they are fooled by the mutating webhook that
instead is not involved into the CRD validation process
on the OLM side.
On the other side, add an additional upgrade lane just for
this is probably too expensive.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=2241327

This is a manual cherry-pick of #2534

Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>