operands/kubevirt: Expand UpdateCr to annotations
Currently the updateCr method is agnostic of changes made on the
Kubevirt CR annotations. This means that HCO cannot add its opinionated
annotations to it.
In order to allow setting annotations to the Kubevirt CR by HCO, and
also allow future json patches to be able to control annotations,
updating the updateCr method to be aware of annotations.
This commit also makes sure that annotation changes are not considered
as "outOfBand" changes, and thus do not trigger any alert.

Signed-off-by: Ram Lavi <ralavi@redhat.com>