Allow modifiying operand annotations using json-patch annotation
PR 2685 sets a new annotation to the KV CR. Currently, the user can't
remove this annotation using the HCO json-patch annotation, in order to
remove this new annotation in case it causes issues in their clusters,
because HCO rejects json-patch annotations if they are trying to modify
fieds that are not under `/spec`.

This PR removes this limitation, allowing the users to modify also
annotations.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>