api: Introduce the AlignCPUs enabler
PR kubevirt/kubevirt#10872 introduced a cluster-wide
setting to enable KubeVirt to request up to two additional
dedicated CPUs in order to complete the total CPU count
to an even parity when using emulator thread isolation.

Introduce a new enabler to HCO's API in order to toggle this
behavior in KubeVirt.

The new API field is optional and defaults to false.

The logic behind it will be introduced in the
following commits.

Signed-off-by: Orel Misan <omisan@redhat.com>
Signed-off-by: Ram Lavi <ralavi@redhat.com>