Introduce AlignCPUs logic
When the FG is enabled ensure the Kubevirt CR will have:
1. The `AlignCPUs` feature gate enabled.
2. The `alpha.kubevirt.io/EmulatorThreadCompleteToEvenParity`
annotation set with a value of an empty string.

Following PR kubevirt/kubevirt#10872, it will enable KubeVirt
to request up to two additional dedicated CPUs in order to complete the
total CPU count to an even parity when using emulator thread isolation.

When the FG is disabled, do the opposite:
1. Make sure the `AlignCPUs` feature gate is disabled.
2. Make sure the above annotation is not set.

The logic introduced in this commit is allowing users to freely
annotate the Kubevirt CR, but HCO will ensure the required
state of the above annotation.

Signed-off-by: Orel Misan <omisan@redhat.com>