[Release-1.10] prepare for next release (#2707)
There is a bug in the publish job, that overrides the existing image
tags, with the same tag.

This PR fix version 1.10.1 by adding a new directory for it, pointing to
the new images.

It also prepare the next version, 1.10.2, by adding a new directory for
it, and modify the publish job to add a new version directory after
publishing.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>