[release 1.11] Bump AAQ to v1.1.4, and rename the component name
AAQ now called "application-aware-quota". One of the configuration field name also renamed.

This PR changes the import path of AAQ to the new path, bumps it to v1.1.4, and rename the component and the
`enableClusterAppsResourceQuota` field to
`allowApplicationAwareClusterResourceQuota`, to match the rename in AAQ.

Also, added a new feature gate, "enableApplicationAwareQuota", to
enable/disable AAQ.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>