[release 1.10] Fix community CSV (#2804)
* Allow running in community hub CI

The community hub CI is trying to install HCO at namespace named
community-kubevirt-hyperconverged. This is not allowed by the operator
code.

This PR adds this namespace to the allowed namesapce list.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>

* Handle the kubevirt-template-validator image in CSV

The kubevirt-template-validator was missing in the related-image list in
the CSV. Also, it was listed with tag instead of with digest, in the SSP
deployment.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>

* add missing annotations to the csv

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>

* set kvci tag to the last working one, 2401211710-a250111

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>

---------

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>