kubvirt: Fix aarch64 OVMF path
As documented [1] downstream aarch64 guests require a different UEFI
firmware to x86_64. This is provided by the edk2-aarch64 package with
the resulting firmware placed under /usr/share/AAVMF.

This change updates the OVMFPath configured for use with such guests on
aarch64 and also updates some tests to assert this is set.

[1] https://access.redhat.com/documentation/en-us/red_hat_enterprise_linux/9/html-single/configuring_and_managing_virtualization/index#how-virtualization-on-arm-64-differs-from-amd64-and-intel64_feature-support-and-limitations-in-rhel-9-virtualization

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>