Deprecate MTQ There are several problems with the MTQ Operator:
Namespace Locking:
- While it functions, bugginess has been observed
  in the wild.

Tightly coupled with KubeVirt Code:
- One-to-one changes are required for updates to
  logic in either repository.

  Modifies the quota set by the cluster admin.

Accounting nuances:
- Both source and target pods are counted until
  migration completes.

The AAQ Operator resolves all these issues;
therefore, the MTQ should be dropped, and we should
fully commit to the AAQ Operator.

Deprecation Plan:

1. Deprecate the MTQ flag, disabling it even
   if activated.

2. Remove MTQ from CSVs and eliminate all
   MTQ-related code from HCO.

This commit accomplishes the first step.
The reason for the two-step removal is the concern
regarding updates with MTQ enabled.

In that case, the MTQ Operator won't be available
to remove MTQ leftovers.

Signed-off-by: bmordeha <bmordeha@redhat.com>