Add the `bump-hco` make target
the new `bump-hco` make target bumps hco to the next version.

By default, the next version is the next minor version, so if the
current version is `1.12.0` or `1.12.3`, the next version will be
`1.13.0`.

It is possible to change this default behavior by setting the
`HCO_BUMP_LEVEL` environment variable. The supported vlaues are `major`,
`minor` (default) or `patch`; e.g.

```shell
HCO_BUMP_LEVEL=patch make bump-hco
```

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>