operands/kubevirt: Set Kubevirt CR new primary UDN Binding plugin
The ability to register network bindings to Kubevirt VMs was proposed
[0] in order to allow Kubevirt VMs to connect to network bindings and
gain advanced user defined network capabilities.
The configuration is added to the kubevirt CR (kubevirt feature gate* +
configuring the binding parameters).

This commit deploys and registers a primary user-defined-network (UDN)
binding to the kubevirt CR, so it could be used by kubevirt users.

* enabling the `NetworkBindingPlugins` kubevirt
feature-gate on kubevirt CR is not needed as it is already set by
default by HCO [1].

[0] https://github.com/kubevirt/community/pull/280
[1]
https://github.com/kubevirt/hyperconverged-cluster-operator/pull/2603

Signed-off-by: Ram Lavi <ralavi@redhat.com>