kubevirt-console-plugin: [release-1.12] Use two replicas in case of highly available cluster (#3035)
In case the cluster has an HighlyAvailable infrastructure, the kubevirt-console-plugin and kubevirt-console-proxy should be highly available as well,
in order to avoid losing the Virtualization plugin when the hosting node is down. By running these pods with two replicas and PodAntiAffinity rule, we
can ensure the plugin will remain available when a node is down.

Signed-off-by: Oren Cohen <ocohen@redhat.com>