Improve the monitoring functional test
The test now reads the kubevirt_hco_out_of_band_modifications_total metric before
modifying the KubeVirt CR. When reading from prometheus, it can take some time for
it to update from the HCO endpoints. That leads to flaky behavior, because the
initial value may be wrong and not up-to-date.

This PR reads the initial value directly from the HCO pod, by adding a temporary
route for the HCO prometheus endpoint, so it is always up-to-date.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>