[release-1.12] dataImportCronTemplates: Remove CentOS 7 & stream 8 (#3066)
* dataImportCronTemplates: Remove CentOS 7 & stream 8

Both releases are now EOL with their associated containerdisks also
deprecated ahead of removal in the near future:

https://github.com/kubevirt/containerdisks/issues/154
https://github.com/kubevirt/containerdisks/issues/152

Conflicts:
  assets/dataImportCronTemplates/dataImportCronTemplates.yaml
  tests/func-tests/golden_image_test.go

NOTE: The conflict is due to commits c70bb07 and 4e8265f missing in
release-1.12.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>
Co-authored-by: Felix Matouschek <fmatouschek.com>

* tests: Remove use of CentOS 8 from golden image func tests

The source of this image stream is also changed to the smaller and more
up to date containerdisks project version of the image.

Conflicts:
  tests/func-tests/golden_image_test.go

NOTE: The conflict is due to commit 4e8265f missing in release-1.12.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>
Co-authored-by: Felix Matouschek <fmatouschek@redhat.com>

---------

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>
Co-authored-by: Lee Yarwood <lyarwood@redhat.com>