[release-1.9] dataImportCronTemplates: Remove CentOS 7 & stream 8 (#3069)
* dataImportCronTemplates: Remove CentOS 7 & stream 8

Both releases are now EOL with their associated containerdisks also
deprecated ahead of removal in the near future:

https://github.com/kubevirt/containerdisks/issues/154
https://github.com/kubevirt/containerdisks/issues/152

Conflicts:
  assets/dataImportCronTemplates/dataImportCronTemplates.yaml
  tests/func-tests/golden_image_test.go
  hack/check_golden_images.sh

NOTE: The conflict is due to commits 2e339b9, ffe5abc
and c70bb07 missing in release-1.9.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>
Co-authored-by: Felix Matouschek <fmatouschek@redhat.com>

* tests: Remove use of CentOS 7 and 8 from func tests

The source of this image stream is also changed to the smaller and more
up to date containerdisks project version of the image.

Conflicts:
  tests/func-tests/golden_image_test.go
  hack/check_golden_images.sh
  tests/func-tests/dict_test.go

NOTE: The conflict is due to commit 2e339b9 missing in release-1.9.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>
Co-authored-by: Felix Matouschek <fmatouschek@redhat.com>

---------

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>
Co-authored-by: Lee Yarwood <lyarwood@redhat.com>