Refactor: Simplify adding operand handler
Allow adding a single handler, in addition to the exsisting adding of
multiple handlers.

The original `addOperands` handlers was desiged for unknown number of
handlers to be initiated on runtime. But for single handlers, there is no
reason to do that.

This PR adds an option to add a single handler for the regular use-case,
when just adding a new known handler.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>