Reduce number of events
HCO triggers too many events. This is creating noise.

This PR limits the number of event triggered by HCO:

Removed the events for failed update of the HyperConverged CR or status

Removed the event about upgrade in progress. The event for getting into
upgrade mode, and exit upgrade mode are enough.

Trigger the "HCO Reconcile completed successfully" event only if HCO
wasn't ready before, and just becomes ready.

Trigger the "Not all the operators are ready" event only if HCO was
ready before, and now becomes not-ready.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>