Bump kubebuilder to go/v4
operator-sdk, from verstion v1.38.0, does not support kubebuilder
go/v4 anymore. Any attempt to run the new operator-sdk in the HCO repo,
fails even if the request is not related to the code, e.g. to install
olm.

This PR bumps the kubebuilder go version to go/v4, to allow running the
newest operator-sdk in the hco repo, e.g. for deploying HCO in tests.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>