Don't rely on InfraStructureTopology for infra HA
There are some cases in which the number of worker nodes are changing throughout the lifecycle of the cluster, and the InfrastructureTopology
in the Infrastructure resource is statically set at cluster installation type and it is not getting updated dynamically.
For example, an SNO cluster that is being added a new worker node. In that case, the infrastructure topology should be updated to
'HighlyAvailable' and don't remain 'SingleReplica'.

Signed-off-by: Oren Cohen <ocohen@redhat.com>