Fix the monitoring test
The "KubeVirtCRModified alert should fired when there is a modification
on a CR" functional test became flacky.

The test changes the KubeVirt CI, and then checks several metrics and
alert. When checking the "kubevirt_hyperconverged_operator_health_status"
alert, the test sometimes fails, as KubeVirt get into non stable state
for a short while for every chenage in the KubeVirt CR, making the
health alret to be in "critical" status, instead of "warning".

This PR fixes the issue by modifiying the CDI CR instead of the KubeVirt
CR, to get the same effect.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>