Remove unused code, that drop CRD version on upgrade
The v1aplpha1 API version of HCO was removed many versions back, and we
can be sure that by now, it's not exist anymore. The code that removes
this version from the CRD on upgrade is not needed.

This commit droped the code that removes the v1alpha1 API version for
the CRD on upgrade.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>