Introduce the crwriter tool
The crwriter tool generates the HyperConverged CR with its default
values, in one of three formats: yaml (default), json, or go file that
contains a json string.

The tool was added in order to auto generate a file that HCO will use to
validate the upgradePatches.json file on boot, even if the
HyperConverged CR is not deployed yet.

It can be used in other places, like replacing the code in the
manifest-templator tool that generates the deploy/hco.cr.yaml, to be
more explicit, or to deploy the HyperConverged CR on hack/deploy.sh.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>