Move the upgradePatches file validation to boot
HCO validates the upgradePatches.json file only after the HyperConverged
CR is created, and if it's not valid, HCO exit with error. This is too
late to kill the HCO pod. The pod should be killed on boot, if something
is not right. The reason for this behavior was that HCO need the
HyperConverged CR in order to run the patches on it.

This commit generates a json file of the HyperConverged CR, using the
new crwriter tool, and then use it to validate the upgradePathes.json
file, without reading the actual CR, but use the pre-generated json
file.

Then, this commit moves the validation code to boot.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>