fix override-bot triggers
The bot now can run from a comment in the PR, from a schedul, or
manually from github ui.

But the action condition requires a PR number to not be empty. That
makes the schedule and the manual triggers to be ignored and skipped.

This commit fixes the condition to allow the other 2 triggers.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>