Stop recalculate namespaced names
The current code re-calculates the expected namespaced name to resolve
reconcile request. But the calculation is based on data that is never
changed.

This commit moves all the reconcile request resolving logic to a
seperate package, for better readability and maintenance. The namespaced
names are only calulated once, at boot time.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>