Bump golangci-lint to v1.64.4
Note: this version can't compiled with go1.23.5, as currently defined in
our go.mod file. This is why we specify the go toolchain in the
Makefile, and set it to v1.23.4

This PR also remove the running golangci-lint in the tests directory,
because it's a left over from the the split of the go.mod in this
directory, that was already canceled.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>