Fix upgrade-issue in AWS
Many upgrade test are failing in AWS, when the VM launcher image is not
replaced after upgrade.

The nodes in AWS does not have the same labels, and KV can't find a node
live-migrate the VM.

This commit disables the VM's "gds-no" CPU feature, so the VM should now
find avaliable node.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>