[upgrade test] set common CPU Model for aws and azure
When performing a live migration, virt-controller picks a target node that has all of the cpu features as the source node.
if the cluster nodes have different CPU features, the VMIM will be stuck at Pending.
setting defaultCPUModel for AWS and Azure to ensure all of the cpu features in all nodes are aligned and live migration will be done.

Signed-off-by: Oren Cohen <ocohen@redhat.com>