Fix issue in upgrade tests
We have an inherent issue in our upgrade tests. The test
is taking the unstable bundles from quay. We build these
bundles each time we merge a PR.

All the images in the unstable bundles are listed with
their digests, rather than their tag, in order to support
disconnected environment.

If this process is failed in the middle, e.g. when the
operator image was already built and pushed to quay.io,
but the bundle image failed, then the bundle is actually
broken: it is an old version of the bundle, pointing to
images that are no longer there, because the operator
images are was already replaced.

This PR fixes this issue by first pushing the images with
a temporary tag, and only if everything went well, retugs
the images with the unstable tag.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>