Add multi-arch annotation to DICT (#3521)
* Add the new enableMultiArchCommonBootImageImport featureGate

enableMultiArchCommonBootImageImport allows the HCO to run on heterogeneous
clusters with different CPU architectures.

Setting this field to true will allow the HCO to create Golden Images for
different CPU architectures.

This feature is in Developer Preview.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>

* Drop annotation from DataImportCronTemplate in SSP

If the enableMultiArchCommonBootImageImport feature gate is "false"
(or missing), do not copy the `"ssp.kubevirt.io/dict.architectures"`
annotation to the SSP CR.

If the feature gate is set to "true, do copy the annotation.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>

* Added the ssp.kubevirt.io/dict.architectures annotation

Added the ssp.kubevirt.io/dict.architectures annotation to the
DataImportCronTemplate objects in
assets/dataImportCronTemplates/dataImportCronTemplates.yaml, by running
`make annotate-dicts`.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>

* update manifests with the new feature gate

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>

---------

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>