Fix the controllers/operands unit tests
Unit tests must be stand alone, pass when run as a single test and must
not be depend on other previous tests.

This is not the case in the controllers/operands package, where
several tests are failing when running individually (e.g. with `FIt`),
or when running tests with random order.

This PR moves all the required settings to the top level of the test
file, and now the the test are all independent.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>