Filter unsupported architectures from DICTs
If there are no supported architectures for a specific DICT, it will not
added to the SSP CR.

The DICTs in the HyperConverged status will have the
`ssp.kubevirt.io/dict.architectures` annotation only with the
architectures that are both suppported by the image, and by the
cluster.

But if there are no supported architectures for a specific DICT, in
addition to not be copied to the SSP CR, its DICT object in the HC
status will have the the new `Deployed` condition in its status,
with status of `False` and reason of `"UnsupportedArchitectures"`.

Also, each DICT object in the HyperConverged status, will have the
new status.originalSupportedArchitectures, with the original values
from the annotation (for debug, troubleshooting and so on).

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>