Filter unsupported architectures from DICTs
If there are no supported architectures for a specific DICT, it will not
added to the SSP CR.

The DICTs in the HyperConverged status will have the
`ssp.kubevirt.io/dict.architectures` annotation only with the
architectures that are both suppported by the image, and by the
cluster.

But if there are no supported architectures for a specific DICT, in
addition to not be copied to the SSP CR, it object in the HC status will
have the `ssp.kubevirt reason of `"UnsupportedArchitectures"`.
original values (for debug), and the new `"Deployed"` condition with
status of `False`, and a reason of

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>