Monitor the cluster architectures (#3594)
* monitor and cache the cluster node architectures

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>

* API changes

1. add the new `nodeInfo` field to the HyperConverged's status
2. add the `conditions` field to the DICT status type
3. add the `originalSupportedArchitectures` field to the DICT status
   type.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>

* Set the new status fields when the nodeinfo is changed

Move the setting of the HA status field to the HC controller.

The nodes controller only triggers the reconciliation of the HC, but
not changes the HC CR status.

Trigger the HC reconciliation will also promise the right setting of the
DataImportCronTemplates in SSP, in a following commit.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>

* validate the nodePlace Affinity field

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>

* collect node information on boot time

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>

* update SSP validation

Use the HyperConverged status new fields, to get the node architectures
for the SSP validation.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>

* Filter unsupported architectures from DICTs

If there are no supported architectures for a specific DICT, it will not
added to the SSP CR.

The DICTs in the HyperConverged status will have the
`ssp.kubevirt.io/dict.architectures` annotation only with the
architectures that are both suppported by the image, and by the
cluster.

But if there are no supported architectures for a specific DICT, in
addition to not be copied to the SSP CR, its DICT object in the HC
status will have the the new `Deployed` condition in its status,
with status of `False` and reason of `"UnsupportedArchitectures"`.

Also, each DICT object in the HyperConverged status, will have the
new status.originalSupportedArchitectures, with the original values
from the annotation (for debug, troubleshooting and so on).

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>

* Remove anused parameter from the NewSSP function

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>

---------

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>