Remove redundant check from the HC predicate (#3611)
In the `HyperConverged` predicate in the node controller, we trigger
reconcile on `HyperConverged` CR update, if the workloads or the infra
field was modified. But we actually don't need to check the infra field,
because it does not affect the node counting, but only the workloads
field does.

This PR drop the unnecessary check of the infra field.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>