Add two new labels for NetworkPolicy (#3569)
Optionally add two new labels to the operator and the webhook pods, in
the CSV.

The `hco.kubevirt.io/allow-access-cluster-services` label is used to
allow egress to the DNS and the API-Server.

The `hco.kubevirt.io/allow-prometheus-access` label is used to allow
ingress to the metrics port of the pod.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>