Secure the metrics endpoint with https
While working on network policies, it appears that network policies do
not work with named ports (e.g. "metrics"), but need the port number of
the pod.

The majority of the pods in KubeVirt project, uses port 8443, so we must
align with them to be able to use the same network policies.

But moving to port 8443 also requires using https instead of http.

This PR changes the protocol of the metrics endpoint to https, and the
port to 8443, both in the operator and the webhook.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>